[win.buckconfig] add external:W0 to cxxflags
Buck passes the include path for python like with /external:I (I think
because the python target is a prebuilt_cxx_library?), which is a
compiler error if /external:Wn is not also passed. I susped this error
is new in VS2019 but have not verified that.

see: https://docs.microsoft.com/en-us/cpp/build/reference/external-external-headers-diagnostics